Added auth for TopBar username downlaod
