+++ logs/api-gateway.log
@@ -173,3 +173,29 @@ d72b0738ca25 | 2025-07-06 21:40:50.584 | info | Gateway ready to handle requests
 4e67afb19dc8 | 2025-07-09 14:48:32.948 | info | Request completed: OPTIONS /api/auth/login -> 200 in 0ms
 4e67afb19dc8 | 2025-07-09 14:48:32.954 | info | Incoming request: POST /api/auth/login from 172.18.0.1
 4e67afb19dc8 | 2025-07-09 14:48:33.035 | info | Request completed: POST /api/auth/login -> 200 in 81ms
+68a77337081d | 2025-08-10 14:29:10.184 | info | Starting API Gateway
+68a77337081d | 2025-08-10 14:29:10.184 | info | Registering AuthService endpoints
+68a77337081d | 2025-08-10 14:29:10.185 | info | AuthService endpoints registered successfully
+68a77337081d | 2025-08-10 14:29:10.185 | info | Registering AuthHello endpoints
+68a77337081d | 2025-08-10 14:29:10.185 | info | AuthHello endpoints registered successfully
+68a77337081d | 2025-08-10 14:29:10.185 | info | API Gateway configured with endpoints:
+68a77337081d | 2025-08-10 14:29:10.186 | info |   POST /api/auth/login
+68a77337081d | 2025-08-10 14:29:10.186 | info |   POST /api/auth/register
+68a77337081d | 2025-08-10 14:29:10.186 | info |   POST /api/auth/refresh
+68a77337081d | 2025-08-10 14:29:10.186 | info |   POST /api/auth/logout
+68a77337081d | 2025-08-10 14:29:10.187 | info |   GET  /api/hello
+68a77337081d | 2025-08-10 14:29:10.187 | info | Starting HTTP server on port :8083
+68a77337081d | 2025-08-10 14:29:10.187 | info | Gateway ready to handle requests
+6340767ccf88 | 2025-08-10 15:06:45.934 | info | Starting API Gateway
+6340767ccf88 | 2025-08-10 15:06:45.936 | info | Registering AuthService endpoints
+6340767ccf88 | 2025-08-10 15:06:45.940 | info | AuthService endpoints registered successfully
+6340767ccf88 | 2025-08-10 15:06:45.948 | info | Registering AuthHello endpoints
+6340767ccf88 | 2025-08-10 15:06:45.961 | info | AuthHello endpoints registered successfully
+6340767ccf88 | 2025-08-10 15:06:45.961 | info | API Gateway configured with endpoints:
+6340767ccf88 | 2025-08-10 15:06:45.961 | info |   POST /api/auth/login
+6340767ccf88 | 2025-08-10 15:06:45.961 | info |   POST /api/auth/register
+6340767ccf88 | 2025-08-10 15:06:45.961 | info |   POST /api/auth/refresh
+6340767ccf88 | 2025-08-10 15:06:45.962 | info |   POST /api/auth/logout
+6340767ccf88 | 2025-08-10 15:06:45.962 | info |   GET  /api/hello
+6340767ccf88 | 2025-08-10 15:06:45.962 | info | Starting HTTP server on port :8083
+6340767ccf88 | 2025-08-10 15:06:45.962 | info | Gateway ready to handle requests
