+++ logs/auth-service.log
@@ -20,3 +20,9 @@ d781fdeab73d | 2025-07-06 23:13:52.872 | info | Auth Service running on port :30
 9fd18cb7a4bd | 2025-07-09 14:47:22.237 | info | Starting Auth Service
 9fd18cb7a4bd | 2025-07-09 14:47:22.256 | info | Database connection established
 9fd18cb7a4bd | 2025-07-09 14:47:22.258 | info | Auth Service running on port :3003
+d86f4dbe80e4 | 2025-08-10 14:29:11.030 | info | Starting Auth Service
+d86f4dbe80e4 | 2025-08-10 14:29:11.037 | info | Database connection established
+d86f4dbe80e4 | 2025-08-10 14:29:11.038 | info | Auth Service running on port :3003
+6160d8cae0be | 2025-08-10 15:06:46.776 | info | Starting Auth Service
+6160d8cae0be | 2025-08-10 15:06:46.792 | info | Database connection established
+6160d8cae0be | 2025-08-10 15:06:46.794 | info | Auth Service running on port :3003
